fhs-xmltv-tools-0.1.3/requirements.txt
py-xmltv
typer[all]
pyyaml
